+++ src/services/events/index.ts
@@ -1,5 +1,5 @@
 import * as Logging from 'hyperview/src/services/logging';
-import { ON_EVENT_DISPATCH } from 'hyperview/src/types';
+import { ON_EVENT_DISPATCH } from './types';
 import { TinyEmitter } from 'tiny-emitter';
 
 const tinyEmitter = new TinyEmitter();
