+++ src/services/events/types.ts
@@ -1,0 +1,1 @@
+export const ON_EVENT_DISPATCH = 'hyperview:on-event';
